refactor w/o test
