+++ cli_tools/gce_export/main.go
@@ -23,15 +23,10 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"math"
-	"math/rand"
 	"net/http"
 	"os"
-	"path"
 	"path/filepath"
-	"regexp"
 	"runtime"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -40,7 +35,6 @@ import (
 	storageutils "github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/storage"
 	"github.com/dustin/go-humanize"
 	gzip "github.com/klauspost/pgzip"
-	"google.golang.org/api/googleapi"
 	"google.golang.org/api/option"
 	htransport "google.golang.org/api/transport/http"
 )
@@ -55,7 +49,6 @@ var (
 	level                    = flag.Int("level", 3, "level of compression from 1-9, 1 being best speed, 9 being best compression")
 	bufferSize               = flag.String("buffer_size", "1GiB", "max buffer size to use")
 	workers                  = flag.Int("workers", runtime.NumCPU(), "number of upload workers to utilize")
-	gcsPermissionErrorRegExp = regexp.MustCompile(".*does not have storage.objects.create access to .*")
 )
 
 // progress is a io.Writer that updates total in Write.
@@ -71,16 +64,6 @@ func (p *progress) Write(b []byte) (int, error) {
 	return len(b), nil
 }
 
-const letters = "bdghjlmnpqrstvwxyz0123456789"
-
-func randString(n int) string {
-	gen := rand.New(rand.NewSource(time.Now().UnixNano()))
-	b := make([]byte, n)
-	for i := range b {
-		b[i] = letters[gen.Int63()%int64(len(letters))]
-	}
-	return string(b)
-}
 
 func splitLicenses(input string) []string {
 	if input == "" {
@@ -91,211 +74,6 @@ func splitLicenses(input string) []string {
 		ls = append(ls, l)
 	}
 	return ls
-}
-
-type bufferedWriter struct {
-	// These fields are read only.
-	cSize    int64
-	prefix   string
-	ctx      context.Context
-	id       string
-	bkt, obj string
-
-	upload    chan string
-	tmpObjs   []string
-	tmpObjsMx sync.Mutex
-
-	sync.Mutex
-	sync.WaitGroup
-	bytes int64
-	part  int
-	file  *os.File
-}
-
-func (b *bufferedWriter) addObj(obj string) {
-	b.tmpObjsMx.Lock()
-	b.tmpObjs = append(b.tmpObjs, obj)
-	b.tmpObjsMx.Unlock()
-}
-
-func (b *bufferedWriter) uploadWorker() {
-	defer b.Done()
-	for in := range b.upload {
-		for i := 1; ; i++ {
-			err := func() error {
-				client, err := gcsClient(b.ctx)
-				if err != nil {
-					return err
-				}
-				defer client.Close()
-
-				file, err := os.Open(in)
-				if err != nil {
-					return err
-				}
-				defer file.Close()
-
-				tmpObj := path.Join(b.obj, strings.TrimPrefix(in, b.prefix))
-				b.addObj(tmpObj)
-				dst := client.Bucket(b.bkt).Object(tmpObj).NewWriter(b.ctx)
-				if _, err := io.Copy(dst, file); err != nil {
-					if io.EOF != err {
-						return err
-					}
-				}
-
-				return dst.Close()
-			}()
-			if err != nil {
-				// Don't retry if permission error as it's not recoverable.
-				gAPIErr, isGAPIErr := err.(*googleapi.Error)
-				if isGAPIErr && gAPIErr.Code == 403 && gcsPermissionErrorRegExp.MatchString(gAPIErr.Message) {
-					fmt.Printf("GCEExport: %v", err)
-					os.Exit(2)
-				}
-
-				fmt.Printf("Failed %v time(s) to upload '%v', error: %v\n", i, in, err)
-				if i > 16 {
-					log.Fatal(err)
-				}
-
-				fmt.Printf("Retrying upload '%v' after %v second(s)...\n", in, i)
-				time.Sleep(time.Duration(1*i) * time.Second)
-				continue
-			}
-			os.Remove(in)
-			break
-		}
-	}
-}
-
-func (b *bufferedWriter) newChunk() error {
-	fp := path.Join(b.prefix, fmt.Sprint(b.id, "_part", b.part))
-	f, err := os.Create(fp)
-	if err != nil {
-		return err
-	}
-
-	b.bytes = 0
-	b.file = f
-	b.part++
-
-	return nil
-}
-
-func (b *bufferedWriter) flush() error {
-	if err := b.file.Close(); err != nil {
-		return err
-	}
-
-	b.upload <- b.file.Name()
-	return nil
-}
-
-func (b *bufferedWriter) Close() error {
-	if err := b.flush(); err != nil {
-		return err
-	}
-	close(b.upload)
-	b.Wait()
-
-	client, err := gcsClient(b.ctx)
-	if err != nil {
-		return err
-	}
-	defer client.Close()
-
-	// Compose the object.
-	for i := 0; ; i++ {
-		var objs []*storage.ObjectHandle
-		// Max 32 components in a single compose.
-		l := math.Min(float64(32), float64(len(b.tmpObjs)))
-		for _, obj := range b.tmpObjs[:int(l)] {
-			objs = append(objs, client.Bucket(b.bkt).Object(obj))
-		}
-		if len(objs) == 1 {
-			if _, err := client.Bucket(b.bkt).Object(b.obj).CopierFrom(objs[0]).Run(b.ctx); err != nil {
-				return err
-			}
-			objs[0].Delete(b.ctx)
-			break
-		}
-		newObj := client.Bucket(b.bkt).Object(path.Join(b.obj, b.id+"_compose_"+strconv.Itoa(i)))
-		b.tmpObjs = append([]string{newObj.ObjectName()}, b.tmpObjs[int(l):]...)
-		if _, err := newObj.ComposerFrom(objs...).Run(b.ctx); err != nil {
-			return err
-		}
-		for _, o := range objs {
-			o.Delete(b.ctx)
-		}
-	}
-	return nil
-}
-
-func (b *bufferedWriter) Write(d []byte) (int, error) {
-	b.Lock()
-	defer b.Unlock()
-
-	if b.file == nil {
-		if err := b.newChunk(); err != nil {
-			return 0, err
-		}
-	}
-
-	b.bytes += int64(len(d))
-	if b.bytes >= b.cSize {
-		if err := b.flush(); err != nil {
-			return 0, err
-		}
-		if err := b.newChunk(); err != nil {
-			return 0, err
-		}
-		b.bytes = int64(len(d))
-	}
-	n, err := b.file.Write(d)
-	if err != nil {
-		return 0, err
-	}
-
-	return n, nil
-}
-
-func gcsClient(ctx context.Context) (*storage.Client, error) {
-	//return storage.NewClient(ctx)
-	baseTransport := &http.Transport{
-		DisableKeepAlives:     false,
-		MaxIdleConns:          0,
-		MaxIdleConnsPerHost:   1000,
-		MaxConnsPerHost:       0,
-		IdleConnTimeout:       60 * time.Second,
-		ResponseHeaderTimeout: 5 * time.Second,
-		TLSHandshakeTimeout:   5 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-	}
-	transport, err := htransport.NewTransport(ctx, baseTransport)
-	if err != nil {
-		return nil, err
-	}
-	return storage.NewClient(ctx, option.WithHTTPClient(&http.Client{Transport: transport}),
-		option.WithCredentialsFile(*oauth))
-}
-
-func newBuffer(ctx context.Context, size, workers int64, prefix, bkt, obj string) *bufferedWriter {
-	b := &bufferedWriter{
-		cSize:  size / workers,
-		prefix: prefix,
-		id:     randString(5),
-
-		upload: make(chan string),
-		bkt:    bkt,
-		obj:    obj,
-		ctx:    ctx,
-	}
-	for i := int64(0); i < workers; i++ {
-		b.Add(1)
-		go b.uploadWorker()
-	}
-	return b
 }
 
 func writeGzipProgress(start time.Time, size int64, rp, wp *progress) {
@@ -327,6 +105,26 @@ func writeGzipProgress(start time.Time, size int64, rp, wp *progress) {
 		oldSince = since
 		time.Sleep(30 * time.Second)
 	}
+}
+
+func gcsClient(ctx context.Context) (*storage.Client, error) {
+	//return storage.NewClient(ctx)
+	baseTransport := &http.Transport{
+		DisableKeepAlives:     false,
+		MaxIdleConns:          0,
+		MaxIdleConnsPerHost:   1000,
+		MaxConnsPerHost:       0,
+		IdleConnTimeout:       60 * time.Second,
+		ResponseHeaderTimeout: 5 * time.Second,
+		TLSHandshakeTimeout:   5 * time.Second,
+		ExpectContinueTimeout: 1 * time.Second,
+	}
+	transport, err := htransport.NewTransport(ctx, baseTransport)
+	if err != nil {
+		return nil, err
+	}
+	return storage.NewClient(ctx, option.WithHTTPClient(&http.Client{Transport: transport}),
+		option.WithCredentialsFile(*oauth))
 }
 
 func gzipDisk(file *os.File, size int64, writer io.WriteCloser) error {
@@ -415,7 +213,7 @@ func stream(ctx context.Context, src *os.File, size int64, prefix, bkt, obj stri
 		if err != nil {
 			return err
 		}
-		buf := newBuffer(ctx, int64(bs), int64(*workers), prefix, bkt, obj)
+		buf := storageutils.NewBuffer(ctx, int64(bs), int64(*workers), *oauth, prefix, bkt, obj)
 
 		fmt.Printf("GCEExport: Using %q as the buffer prefix, %s as the buffer size, and %d as the number of workers.\n", prefix, humanize.IBytes(bs), *workers)
 		return gzipDisk(src, size, buf)
